requirements.txt
sentence-transformers
google-generativeai
chromadb
pandas
numpy
scikit-learn
streamlit
python-dotenv